refactor: align deployment detail action ownership
