+++ web/features/deployments/detail/access-tab/developer-api/create-api-key-dialog.tsx
@@ -1,10 +1,7 @@
 'use client'
 
 import type { Environment } from '@dify/contracts/enterprise/types.gen'
-import type { ButtonProps } from '@langgenius/dify-ui/button'
-import type { ReactNode } from 'react'
 import { Button } from '@langgenius/dify-ui/button'
-import { cn } from '@langgenius/dify-ui/cn'
 import {
   Dialog,
   DialogCloseButton,
@@ -26,11 +23,9 @@ import {
 } from '@langgenius/dify-ui/select'
 import { toast } from '@langgenius/dify-ui/toast'
 import { useMutation } from '@tanstack/react-query'
-import { useAtomValue } from 'jotai'
 import { useRef, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { consoleQuery } from '@/service/client'
-import { deploymentRouteAppInstanceIdAtom } from '../../../route-state'
 import { generateApiTokenName } from './api-token-name'
 
 type CreateApiKeyFormValues = {
@@ -38,7 +33,7 @@ type CreateApiKeyFormValues = {
   environmentId: string
 }
 
-function CreateApiKeyDialog({
+export function CreateApiKeyDialog({
   appInstanceId,
   environments,
   open,
@@ -225,62 +220,6 @@ function CreateApiKeyDialogContent({
           </Button>
         </div>
       </Form>
-    </>
-  )
-}
-
-export function ApiKeyGenerateMenu({
-  environments,
-  onCreatedToken,
-  triggerVariant = 'secondary',
-  triggerClassName,
-  children,
-}: {
-  environments: Environment[]
-  onCreatedToken: (token: string) => void
-  triggerVariant?: ButtonProps['variant']
-  triggerClassName?: string
-  children?: (props: { trigger: ReactNode }) => ReactNode
-}) {
-  const { t } = useTranslation('deployments')
-  const appInstanceId = useAtomValue(deploymentRouteAppInstanceIdAtom)
-  const [createDialogOpen, setCreateDialogOpen] = useState(false)
-  const [createDialogSessionKey, setCreateDialogSessionKey] = useState(0)
-  const disabled = !appInstanceId || environments.length === 0
-
-  function handleOpenCreateDialog() {
-    const firstEnvironment = environments[0]
-    if (!firstEnvironment)
-      return
-
-    setCreateDialogSessionKey(sessionKey => sessionKey + 1)
-    setCreateDialogOpen(true)
-  }
-
-  const trigger = (
-    <Button
-      type="button"
-      variant={triggerVariant}
-      disabled={disabled}
-      onClick={handleOpenCreateDialog}
-      className={cn('gap-1.5', triggerClassName)}
-    >
-      <span className="i-ri-add-line size-4" aria-hidden="true" />
-      {t('access.api.newKey')}
-    </Button>
-  )
-
-  return (
-    <>
-      {children ? children({ trigger }) : trigger}
-      <CreateApiKeyDialog
-        appInstanceId={appInstanceId}
-        environments={environments}
-        open={createDialogOpen}
-        sessionKey={createDialogSessionKey}
-        onCreatedToken={onCreatedToken}
-        onOpenChange={setCreateDialogOpen}
-      />
     </>
   )
 }
