+++ web/features/deployments/detail/access-tab/developer-api/section.tsx
@@ -1,24 +1,20 @@
 'use client'
 
 import type {
-  AccessChannels,
   ApiKey,
   Environment,
 } from '@dify/contracts/enterprise/types.gen'
 import type { ReactNode } from 'react'
 import { Button } from '@langgenius/dify-ui/button'
-import { Switch, SwitchSkeleton } from '@langgenius/dify-ui/switch'
-import { useMutation } from '@tanstack/react-query'
 import { useAtomValue } from 'jotai'
 import { useState } from 'react'
 import { useTranslation } from 'react-i18next'
-import { consoleQuery } from '@/service/client'
 import { DeploymentEmptyState, DeploymentStateMessage } from '../../../components/empty-state'
 import { deploymentRouteAppInstanceIdAtom } from '../../../route-state'
 import { CopyPill } from '../components/endpoint'
 import { developerApiSettingsQueryAtom } from '../state'
-import { ApiKeyGenerateMenu } from './api-key-generate-menu'
 import { ApiKeyList } from './api-key-list'
+import { CreateApiKeyAction } from './create-api-key-action'
 import { CreatedApiTokenDialog } from './created-token-dialog'
 import { DeveloperApiDocsDrawer } from './docs-drawer'
 import { DeveloperApiSkeleton } from './skeleton'
@@ -26,61 +22,6 @@ import { DeveloperApiSkeleton } from './skeleton'
 type CreatedApiToken = {
   appInstanceId: string
   token: string
-}
-
-function DeveloperApiSwitch({ checked, accessChannels, disabled }: {
-  checked: boolean
-  accessChannels?: AccessChannels
-  disabled?: boolean
-}) {
-  const { t } = useTranslation('deployments')
-  const appInstanceId = useAtomValue(deploymentRouteAppInstanceIdAtom)
-  const toggleDeveloperAPI = useMutation(consoleQuery.enterprise.accessService.updateAccessChannels.mutationOptions())
-
-  return (
-    <Switch
-      aria-label={t('access.api.developerTitle')}
-      checked={checked}
-      disabled={disabled || !appInstanceId}
-      loading={toggleDeveloperAPI.isPending}
-      onCheckedChange={(enabled) => {
-        if (!appInstanceId)
-          return
-
-        toggleDeveloperAPI.mutate({
-          params: { appInstanceId },
-          body: {
-            appInstanceId,
-            webAppEnabled: accessChannels?.webAppEnabled ?? false,
-            developerApiEnabled: enabled,
-          },
-        })
-      }}
-    />
-  )
-}
-
-export function DeveloperApiHeaderSwitch() {
-  const { t } = useTranslation('deployments')
-  const developerApiSettingsQuery = useAtomValue(developerApiSettingsQueryAtom)
-  const accessChannels = developerApiSettingsQuery.data?.accessChannels
-  const apiEnabled = accessChannels?.developerApiEnabled ?? false
-
-  if (developerApiSettingsQuery.isLoading)
-    return <SwitchSkeleton />
-
-  return (
-    <div className="flex items-center gap-2">
-      <span className="system-xs-medium text-text-tertiary">
-        {apiEnabled ? t('overview.enabled') : t('overview.disabled')}
-      </span>
-      <DeveloperApiSwitch
-        checked={apiEnabled}
-        accessChannels={accessChannels}
-        disabled={developerApiSettingsQuery.isError}
-      />
-    </div>
-  )
 }
 
 function ApiKeyListSection({ apiKeys, environments, action }: {
@@ -182,7 +123,7 @@ export function DeveloperApiSection() {
       )}
       {hasSelectableEnvironment
         ? (
-            <ApiKeyGenerateMenu
+            <CreateApiKeyAction
               environments={environments}
               triggerVariant="primary"
               onCreatedToken={token => setCreatedApiToken({ appInstanceId, token })}
@@ -204,7 +145,7 @@ export function DeveloperApiSection() {
                       action={trigger}
                     />
                   )}
-            </ApiKeyGenerateMenu>
+            </CreateApiKeyAction>
           )
         : apiKeys.length === 0
           ? (
