+++ web/features/deployments/detail/index.tsx
@@ -10,7 +10,7 @@ import Link from '@/next/link'
 import { useSelectedLayoutSegment } from '@/next/navigation'
 import { CreateReleaseControl } from '../create-release'
 import { deploymentRouteAppInstanceIdAtom } from '../route-state'
-import { DeveloperApiHeaderSwitch } from './access-tab/developer-api/section'
+import { DeveloperApiHeaderSwitch } from './access-tab/developer-api/developer-api-header-switch'
 import { NewDeploymentHeaderAction } from './deploy-tab/new-deployment-button'
 import { INSTANCE_DETAIL_TAB_KEYS, isInstanceDetailTabKey } from './tabs'
 import { versionsTabLocalAtoms } from './versions-tab/state'
